+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 All notable changes to this project will be documented in this file.
 
+## [0.3.0] - 2025-10-31 - Prime and Function Expansions - [ONGOING]
+
+- Added prime functions:
+  - `isCoprime(a, b)` – Checks whether two integers are coprime (i.e., gcd(a, b) = 1)
+
 ## [0.2.0] - 2025-10-30 - Arithmetic, Combinatorics and Prime Expansion
 
 - Added arithmetic functions: